move viewmodel and start separate conf
