+++ flutter/assets/config/realm.json
@@ -1,0 +1,4 @@
+{
+  "appId": "todo-sync-msgyz",
+  "baseUrl": "https://realm.mongodb.com"
+}
