+++ flutter/pubspec.yaml
@@ -35,9 +35,9 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  realm:
-    path: ./realm_packages/realm/
+  realm: 0.3.0-beta
   flutter_slidable: ^1.2.0
+  path_provider: ^2.0.10
 
 dev_dependencies:
   flutter_test:
@@ -60,10 +60,8 @@ flutter:
   # the material Icons class.
   uses-material-design: true
 
-  # To add assets to your application, add an assets section, like this:
-  # assets:
-  #   - images/a_dot_burr.jpeg
-  #   - images/a_dot_ham.jpeg
+  assets:
+    - assets/config/
 
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
